pyproject.toml
[project]
name = "schwab-mcp"
version = "0.1.0"
description = "Read-only FastMCP server for Charles Schwab brokerage data"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "schwab-py>=1.0",
    "httpx>=0.25.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[tool.setuptools]
py-modules = ["server", "auth", "config", "models"]
packages = ["tools", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
